v0.82.0: 新增factor_store_v5、tail_v5_pipeline模块，更新并行计算及流水线相关代码
新增:
- src/factor_store_v5.rs: 新版因子存储模块
- src/tail_v5_pipeline.rs: v5尾部流水线处理
- examples/: 示例代码

修改:
- Cargo.lock/Cargo.toml: 依赖更新
- src/lib.rs: 注册新模块
- src/backup_reader.rs: 备份读取器更新
- src/bin/worker_pipeline.rs: worker流水线更新
- src/factor_pipeline.rs: 因子流水线优化
- src/fast_csv_reader.rs: CSV读取器优化
- src/features.rs: 特征模块更新
- src/observable_order_metrics.rs: 订单指标更新
- src/order_pair_metrics_pipeline.rs: 订单对指标流水线
- src/parallel_computing.rs: 并行计算优化
- src/parallel_computing_date_only.rs: 日期并行计算
- *.pyi: 类型声明同步更新

+++ src/parallel_computing.rs
@@ -1374,7 +1374,7 @@ fn run_persistent_task_worker(
                         date: task.date,
                         code: task.code,
                         timestamp: chrono::Utc::now().timestamp_millis(),
-                        facs: vec![f64::NAN; expected_result_length],
+                        facs: vec![f32::NAN; expected_result_length],
                     };
 
                     if let Err(e) = result_sender.send(error_result) {
